Refactored db and added AbstractCollectionTest.
