--- src/main/code/de/tu_chemnitz/se/exercise/OutputDevice.kt
@@ -1,8 +1,0 @@
-package de.tu_chemnitz.se.exercise
-
-interface OutputDevice {
-  fun print(message: String)
-
-  @JvmDefault
-  fun printLine(message: String) = print("$message\n")
-}
