--- src/main/code/de/tu_chemnitz/se/exercise/persist/IPersist.kt
@@ -1,5 +1,0 @@
-package de.tu_chemnitz.se.exercise.persist
-
-interface IPersist {
- //persist interfaces as a union
-}
