--- src/main/code/de/tu_chemnitz/se/exercise/persist/configs/ZoomMapsConfig.kt
@@ -1,10 +1,0 @@
-package de.tu_chemnitz.se.exercise.persist.configs
-
-import org.litote.kmongo.Id
-import org.litote.kmongo.newId
-
-data class ZoomMapsConfig(
-  override val _id: Id<ZoomMapsConfig> = newId(),
-  val zoomSpeed : Float,
-  val keyBindings : Set<String>
-) : IConfig
