--- src/main/code/de/tu_chemnitz/se/exercise/persist/data/CodeChartsData.kt
@@ -1,8 +1,0 @@
-package de.tu_chemnitz.se.exercise.persist.data
-
-import org.litote.kmongo.Id
-
-data class CodeChartsData(
-  override val _id: Id<out IData>,
-  val dummyData: String
-) : IData
