--- src/main/code/de/tu_chemnitz/se/exercise/persist/data/UserData.kt
@@ -1,10 +1,0 @@
-package de.tu_chemnitz.se.exercise.persist.data
-
-import org.litote.kmongo.Id
-
-data class UserData(
-  override val _id: Id<out IData>,
-  val firstName: String,
-  val surName: String,
-  val age: Int
-) : IData
