--- src/main/code/de/tu_chemnitz/se/exercise/persist/data/ZoomMapsData.kt
@@ -1,8 +1,0 @@
-package de.tu_chemnitz.se.exercise.persist.data
-
-import org.litote.kmongo.Id
-
-data class ZoomMapsData(
-  override val _id: Id<out IData>,
-  val dummyData: String
-) : IData
