+++ src/main/code/de/tuchemnitz/se/exercise/persist/IPersist.kt
@@ -1,0 +1,5 @@
+package de.tuchemnitz.se.exercise.persist
+
+interface IPersist {
+ //persist interfaces as a union
+}
