+++ src/main/code/de/tuchemnitz/se/exercise/persist/configs/EyeTrackingConfig.kt
@@ -1,0 +1,9 @@
+package de.tuchemnitz.se.exercise.persist.configs
+
+import org.litote.kmongo.Id
+
+//TODO:
+data class EyeTrackingConfig(
+ override val _id: Id<out IConfig>,
+ val dummyVal: String
+) : IConfig
