+++ src/main/code/de/tuchemnitz/se/exercise/persist/data/EyeTrackingData.kt
@@ -1,0 +1,8 @@
+package de.tuchemnitz.se.exercise.persist.data
+
+import org.litote.kmongo.Id
+
+data class EyeTrackingData(
+  override val _id: Id<out IData>,
+  val dummyData: String
+) : IData
