+++ src/test/code/de/tuchemnitz/se/exercise/core/ToolControllerTest.kt
@@ -1,0 +1,23 @@
+package de.tuchemnitz.se.exercise.core
+
+import org.junit.jupiter.api.Test
+
+class ToolControllerTest {
+ // private val controller = ToolController()
+
+  @Test
+  fun `start tool should successfully start tool with config manager`(){
+
+  }
+
+  @Test
+  fun `end tool should successfully close tool and application`(){
+
+  }
+
+  @Test
+  fun `a new user should be successfully created and added to db`(){
+
+  }
+
+}
