+++ src/test/code/de/tuchemnitz/se/exercise/core/configmanager/ConfigManagerTest.kt
@@ -1,0 +1,42 @@
+package de.tuchemnitz.se.exercise.core.configmanager
+
+import org.junit.jupiter.api.Test
+
+class ConfigManagerTest {
+  //private val manager = ConfigManager()
+
+  @Test
+  fun `comparing db and file content should work`(){
+
+  }
+
+  @Test
+  fun `writing in file should work`(){
+
+  }
+
+  @Test
+  fun `reading out of file should work`(){
+
+  }
+
+  @Test
+  fun `assembling all database configs should work`(){
+
+  }
+
+  @Test
+  fun `invoking of get function of a config out of db by id should work`(){
+
+  }
+
+  @Test
+  fun `invoking of save function of a config into the db should work`(){
+
+  }
+
+  @Test
+  fun `setting config path should work`(){
+
+  }
+}
